Name/pyproject.toml
[project]
name = "my_django_project"
version = "0.1.0"
description = "A Django web application"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "Django>=5",
    "djangorestframework",
    "python-dotenv",
    "mysqlclient",
    'mysql-connector-python',
]

dev = [
    "ruff",
    "black",
    "pytest",
    "pytest-django",
]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py310"]

 
